Separated map wrapper object.
