+++ ClientBlazorDemo/Pages/Maps.razor
@@ -1,10 +1,13 @@
 ﻿@page "/maps"
 @using GoogleMapsComponents
 @using GoogleMapsComponents.Maps
+@using System.Diagnostics
+
+@inject IJSRuntime JsRuntime
 
 <h1>Maps</h1>
 
-<GoogleMap ref="@map1" Id="map1" Options="@mapOptions"></GoogleMap>
+<div ref="@map1ElementRef" Id="map1" style="height: 500px"></div>
 <GoogleMap ref="@map2" Id="map2" Options="@mapOptions2"></GoogleMap>
 
 <button onclick="@SyncBounds">Sync map bounds</button>
@@ -15,11 +18,13 @@
 <button onclick="@ZoomIn">Zoom in</button>
 
 @functions {
-    private GoogleMap map1;
+    private Map map1;
     private GoogleMap map2;
 
     private MapOptions mapOptions;
     private MapOptions mapOptions2;
+
+    private ElementRef map1ElementRef;
 
     protected override void OnInit()
     {
@@ -48,39 +53,54 @@
         };
     }
 
+    protected override async Task OnAfterRenderAsync()
+    {
+        if(map1 == null)
+        {
+            map1 = await Map.CreateAsync(JsRuntime, map1ElementRef, mapOptions);
+
+            Debug.WriteLine("Init finished");
+        }
+
+        //if (OnAfterInit != null)
+        //{
+        //    OnAfterInit();
+        //}
+    }
+
     private async Task SyncBounds()
     {
         var bounds = await map1.GetBounds();
-        await map2.FitBounds(bounds);
+        await map2.InteropObject.FitBounds(bounds);
     }
 
     private async Task SyncMapCenter()
     {
         var latLng = await map1.GetCenter();
-        await map2.SetCenter(latLng);
+        await map2.InteropObject.SetCenter(latLng);
     }
 
     private async Task SetMapHeading()
     {
-        await map2.SetHeading(await map2.GetHeading() + 90);
-        var heading = await map2.GetHeading();
+        await map2.InteropObject.SetHeading(await map2.InteropObject.GetHeading() + 90);
+        var heading = await map2.InteropObject.GetHeading();
 
         Console.WriteLine($"Map2 heading is {heading}");
     }
 
     private async Task ToggleTilt()
     {
-        var tilt = await map2.GetTilt();
+        var tilt = await map2.InteropObject.GetTilt();
 
         Console.WriteLine($"Map2 tilt is {tilt}");
 
         if (tilt > 0)
         {
-            await map2.SetTilt(0);
+            await map2.InteropObject.SetTilt(0);
         }
         else
         {
-            await map2.SetTilt(45);
+            await map2.InteropObject.SetTilt(45);
         }
     }
