+++ GoogleMapsComponents/JsObjectRefConverter.cs
@@ -1,4 +1,5 @@
-﻿using Newtonsoft.Json;
+﻿using GoogleMapsComponents.Maps;
+using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -6,14 +7,15 @@ using System.Threading.Tasks;
 
 namespace GoogleMapsComponents
 {
-    internal class MapComponentConverter : JsonConverter<MapComponent>
+    internal class JsObjectRefConverter<T> : JsonConverter<T>
+        where T : IJsObjectRef
     {
-        public override void WriteJson(JsonWriter writer, MapComponent value, JsonSerializer serializer)
+        public override void WriteJson(JsonWriter writer, T value, JsonSerializer serializer)
         {
             serializer.Serialize(writer, new JsObjectRef1(value.Guid));
         }
 
-        public override MapComponent ReadJson(JsonReader reader, Type objectType, MapComponent existingValue, bool hasExistingValue, JsonSerializer serializer)
+        public override T ReadJson(JsonReader reader, Type objectType, T existingValue, bool hasExistingValue, JsonSerializer serializer)
         {
             throw new NotImplementedException();
