+++ GoogleMapsComponents/Maps/MapEventListener.cs
@@ -6,6 +6,10 @@ using System.Threading.Tasks;
 
 namespace GoogleMapsComponents.Maps
 {
+    /// <summary>
+    /// google.maps.MapsEventListener interface
+    /// An event listener, created by google.maps.event.addListener() and friends.
+    /// </summary>
     public class MapEventListener : IDisposable
     {
         private readonly JsObjectRef _jsObjectRef;
