+++ GoogleMapsComponents/Maps/PolylineOptions.cs
@@ -44,8 +44,8 @@ namespace GoogleMapsComponents.Maps
         /// <summary>
         /// Map on which to display Polyline.
         /// </summary>
-        [JsonConverter(typeof(MapComponentConverter))]
-        public MapComponent Map { get; set; }
+        [JsonConverter(typeof(JsObjectRefConverter<Map>))]
+        public Map Map { get; set; }
 
         /// <summary>
         /// The ordered sequence of coordinates of the Polyline.
