+++ GoogleMapsComponents/Maps/RectangleOptions.cs
@@ -45,8 +45,8 @@ namespace GoogleMapsComponents.Maps
         /// <summary>
         /// Map on which to display Rectangle.
         /// </summary>
-        [JsonConverter(typeof(MapComponentConverter))]
-        public MapComponent Map { get; set; }
+        [JsonConverter(typeof(JsObjectRefConverter<Map>))]
+        public Map Map { get; set; }
 
         /// <summary>
         /// The stroke color. All CSS3 colors are supported except for extended named colors.
